+++ packages/typespec-go/test/azure-http-specs/client/naminggroup/naming_client_test.go
@@ -70,9 +70,10 @@ func TestNamingClient_Response(t *testing.T) {
 	client, err := naminggroup.NewNamingClientWithNoCredential("http://localhost:3000", nil)
 	require.NoError(t, err)
 	resp, err := client.Response(context.Background(), nil)
+	_ = resp
 	require.NoError(t, err)
-	require.NotNil(t, resp.DefaultName)
-	require.EqualValues(t, "true", *resp.DefaultName)
+	require.NotNil(t, resp.ClientName)
+	require.EqualValues(t, "true", *resp.ClientName)
 }
 
 func TestModelClient_Client(t *testing.T) {